add job route and move job form to it's own component
