+++ src/pages/index.tsx
@@ -16,7 +16,7 @@ const Home: NextPage = () => {
     <>
       <Head>
         <title>ReactRoles</title>
-        <meta name="description" content="Generated by create-t3-app" />
+        <meta name="description" content="Curated react jobs" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
       <main>
@@ -28,104 +28,6 @@ const Home: NextPage = () => {
 };
 
 export default Home;
-
-const PostJobForm: React.FC = () => {
-  const { refetch: refetchJobs } = api.jobs.getAll.useQuery({});
-  const defaultFormState = {
-    company: "",
-    title: "",
-    salary: 0,
-    description: "",
-    location: "",
-    requirements: "",
-    remote: false,
-  }
-  const [formState, setFormState] = useState(defaultFormState)
-
-  const createJob = api.jobs.create.useMutation({
-    onSuccess: () => {
-      void refetchJobs();
-      setFormState(defaultFormState)
-    },
-  });
-  return (
-    <>
-      <input type="checkbox" id="job-modal" className="modal-toggle" />
-      <div className="modal">
-        <div className="modal-box relative">
-          <label htmlFor="job-modal" className="btn btn-sm btn-circle absolute right-2 top-2">
-            ✕
-          </label>
-          <h3 className="text-lg font-bold">Job</h3>
-          <div className="form-control w-full">
-            <label className="label">
-              <span className="label-text">Company</span>
-            </label>
-            <input value={formState.company} onChange={e => setFormState({ ...formState, company: e.target.value })} type="text" placeholder="Type here" className="input input-bordered w-full outline-none" />
-          </div>
-          <div className="form-control w-full">
-            <label className="label">
-              <span className="label-text">Job title</span>
-            </label>
-            <input value={formState.title} onChange={e => setFormState({ ...formState, title: e.target.value })} type="text" placeholder="Type here" className="input input-bordered w-full outline-none" />
-          </div>
-          <div className="form-control w-full">
-            <label className="label">
-              <span className="label-text">Salary</span>
-            </label>
-            <input value={formState.salary} onChange={e => setFormState({ ...formState, salary: +e.target.value })} type="number" placeholder="Type here" className="input input-bordered w-full outline-none" />
-          </div>
-          <div className="form-control w-full">
-            <label className="label">
-              <span className="label-text">Job description?</span>
-            </label>
-            <input value={formState.description} onChange={e => setFormState({ ...formState, description: e.target.value })} type="text" placeholder="Type here" className="input input-bordered w-full outline-none" />
-          </div>
-          <div className="form-control w-full">
-            <label className="label">
-              <span className="label-text">Location</span>
-            </label>
-            <input value={formState.location} onChange={e => setFormState({ ...formState, location: e.target.value })} type="text" placeholder="Type here" className="input input-bordered w-full outline-none" />
-          </div>
-          <div className="form-control">
-            <label className="label">
-              <span className="label-text">Requirements</span>
-            </label>
-            <textarea value={formState.requirements} onChange={e => setFormState({ ...formState, requirements: e.target.value })} className="textarea textarea-bordered h-24 outline-none" placeholder="Job requirements"></textarea>
-          </div>
-          <div className="form-control">
-            <label className="label cursor-pointer">
-              <span className="label-text">Remote?</span>
-              <input
-                type="checkbox"
-                className="toggle"
-                checked={formState.remote}
-                onChange={e => {
-                  const isChecked = e.target.checked;
-                  setFormState((prevState) => ({
-                    ...prevState,
-                    remote: isChecked,
-                  }));
-                }} />
-            </label>
-          </div>
-          <button
-            className="btn-primary p-2 w-full mt-2"
-            onClick={() => {
-              const jobModal = document.getElementById('job-modal') as HTMLInputElement;
-              if (jobModal) {
-                jobModal.checked = false;
-              }
-              createJob.mutate(formState)
-            }}
-          >
-            Post job
-          </button>
-        </div>
-      </div>
-    </>
-  )
-}
 
 const Content: React.FC = () => {
   const [search, setSearch] = useState<string>('');
@@ -148,14 +50,10 @@ const Content: React.FC = () => {
     <div className="p-16 flex flex-col align-center items-center w-screen justify-center">
       <h1 className="mb-4 text-4xl font-extrabold leading-none tracking-tight text-slate-900 md:text-5xl lg:text-6xl dark:text-white">React Roles</h1>
       <p className="mb-6 text-lg font-normal text-slate-500 lg:text-xl sm:px-16 xl:px-48 dark:text-slate-400 text-center">Welcome to ReactRoles, the premier job board dedicated to React developers! Discover a curated selection of exciting job opportunities in the React ecosystem, ranging from startups to leading tech companies.</p>
-      {/* The button to open modal */}
       {sessionData?.user && (
-        <>
-          <label htmlFor="job-modal" className="btn-primary p-4 rounded">
-            post a job
-          </label>
-          <PostJobForm />
-        </>
+        <button className="btn-primary p-4 rounded">
+          Post a job
+        </button>
       )}
       <div className="my-8 flex justify-between items-center w-full gap-4 max-w-4xl">
         <div className="form-control w-full max-w-2xl">
